+++ src/live/protocol.rs
@@ -18,7 +18,7 @@ use tracing::{debug, error, instrument};
 
 use crate::{ApiKey, Error, USER_AGENT};
 
-use super::Subscription;
+use super::{SlowReaderBehavior, Subscription};
 
 /// Returns the host and port for the live gateway for the given dataset.
 ///
@@ -223,6 +223,8 @@ pub struct SessionOptions<'a> {
     pub heartbeat_interval_s: Option<i64>,
     /// Extension string to append to the user agent.
     pub user_agent_ext: Option<&'a str>,
+    /// The behavior of the gateway when the client falls behind real time.
+    pub slow_reader_behavior: Option<SlowReaderBehavior>,
 }
 
 impl Default for SessionOptions<'_> {
@@ -232,6 +234,7 @@ impl Default for SessionOptions<'_> {
             send_ts_out: false,
             heartbeat_interval_s: None,
             user_agent_ext: None,
+            slow_reader_behavior: None,
         }
     }
 }
@@ -284,6 +287,9 @@ impl AuthRequest {
         );
         if let Some(heartbeat_interval_s) = options.heartbeat_interval_s {
             req = format!("{req}|heartbeat_interval_s={heartbeat_interval_s}");
+        }
+        if let Some(slow_reader_behavior) = options.slow_reader_behavior {
+            req = format!("{req}|slow_reader_behavior={slow_reader_behavior}");
         }
         req.push('\n');
         Self(req)